+++ modules/snooker/src/main/ts/ephox/snooker/api/TableResize.ts
@@ -42,7 +42,6 @@ export interface TableResize {
   readonly refreshBars: (table: SugarElement<HTMLTableElement>) => void;
   readonly hideBars: () => void;
   readonly showBars: () => void;
-  readonly setHoverRefresh: (state: boolean) => void;
   readonly destroy: () => void;
   readonly events: TableResizeEventRegistry;
 }
@@ -86,7 +85,6 @@ const create = (wire: ResizeWire, resizing: ResizeBehaviour, lazySizing: (elemen
     refreshBars: manager.refresh,
     hideBars: manager.hideBars,
     showBars: manager.showBars,
-    setHoverRefresh: manager.setHoverRefresh,
     destroy: manager.destroy,
     events: events.registry
   };