+++ modules/snooker/src/main/ts/ephox/snooker/resize/BarManager.ts
@@ -41,7 +41,6 @@ export interface BarManager {
   readonly off: () => void;
   readonly hideBars: () => void;
   readonly showBars: () => void;
-  readonly setHoverRefresh: (state: boolean) => void;
   readonly events: DragAdjustEvents['registry'];
 }
 
@@ -50,7 +49,6 @@ const resizeBarDragging = Styles.resolve('resizer-bar-dragging');
 export const BarManager = (wire: ResizeWire): BarManager => {
   const mutation = BarMutation();
   const resizing = Dragger.transform(mutation, {});
-  let hoverRefresh = true;
 
   let hoverTable = Optional.none<SugarElement<HTMLTableElement>>();
 
@@ -75,10 +73,6 @@ export const BarManager = (wire: ResizeWire): BarManager => {
     const newX = CellUtils.getCssValue(target, dir);
     const oldX = CellUtils.getAttrValue(target, 'data-initial-' + dir, 0);
     return newX - oldX;
-  };
-
-  const setHoverRefresh = (state: boolean) => {
-    hoverRefresh = state;
   };
 
   /* Resize the column once the user releases the mouse */
@@ -144,7 +138,7 @@ export const BarManager = (wire: ResizeWire): BarManager => {
         }
       },
       (table) => {
-        if (hoverRefresh) {
+        if (resizing.isActive()) {
           hoverTable = Optional.some(table);
           Bars.refresh(wire, table);
         }
@@ -176,7 +170,6 @@ export const BarManager = (wire: ResizeWire): BarManager => {
     off: resizing.off,
     hideBars: Fun.curry(Bars.hide, wire),
     showBars: Fun.curry(Bars.show, wire),
-    setHoverRefresh,
     events: events.registry
   };
 };